products/serializers.py

from rest_framework import serializers
from .models import Brand, Product

class BrandSerializer(serializers.ModelSerializer):
    class Meta:
        model = Brand
        fields = ['id', 'name']

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ['id', 'name', 'asin', 'sku', 'image', 'brand']
